[ottf] consolidate OTTF console functions in the console library
